// spa/src/backend-api/apis/UploadApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CloudNativeQuickShare API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import * as runtime from '../runtime';
import type {
  ErrorResponse,
  UploadFileResponse,
} from '../models/index';
import {
    ErrorResponseFromJSON,
    ErrorResponseToJSON,
    UploadFileResponseFromJSON,
    UploadFileResponseToJSON,
} from '../models/index';

export interface UploadFileEndpointRequest {
    file?: Blob;
    autoShare?: boolean;
}

/**
 * 
 */
export class UploadApi extends runtime.BaseAPI {

    /**
     */
    async uploadFileEndpointRaw(requestParameters: UploadFileEndpointRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<UploadFileResponse>> {
        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        const consumes: runtime.Consume[] = [
            { contentType: 'multipart/form-data' },
        ];
        // @ts-ignore: canConsumeForm may be unused
        const canConsumeForm = runtime.canConsumeForm(consumes);

        let formParams: { append(param: string, value: any): any };
        let useForm = false;
        // use FormData to transmit files using content-type "multipart/form-data"
        useForm = canConsumeForm;
        if (useForm) {
            formParams = new FormData();
        } else {
            formParams = new URLSearchParams();
        }

        if (requestParameters['file'] != null) {
            formParams.append('file', requestParameters['file'] as any);
        }

        if (requestParameters['autoShare'] != null) {
            formParams.append('autoShare', requestParameters['autoShare'] as any);
        }

        const response = await this.request({
            path: `/api/upload`,
            method: 'POST',
            headers: headerParameters,
            query: queryParameters,
            body: formParams,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => UploadFileResponseFromJSON(jsonValue));
    }

    /**
     */
    async uploadFileEndpoint(requestParameters: UploadFileEndpointRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<UploadFileResponse> {
        const response = await this.uploadFileEndpointRaw(requestParameters, initOverrides);
        return await response.value();
    }

}